Adjusted multiple views
