+++ irs-scout/src/Client/java/com/stuin/irs_scout/Updater.java
@@ -31,13 +31,15 @@ public class Updater {
 
         @Override
         public void onFinish() {
-            class Status extends Request {
-                @Override
-                public void run(List<String> s) {
-                    if(!s.get(0).equals((String.valueOf(matchMaker.match.number)))) matchMaker.newMatch();
+            if(status.getText().equals("Waiting")) {
+                class Status extends Request {
+                    @Override
+                    public void run(List<String> s) {
+                        if(!s.get(0).equals(String.valueOf(matchMaker.match.number))) matchMaker.newMatch();
+                    }
                 }
-            }
-            new Status().start("/tablet?status=" + status.getText().toString());
+                new Status().start("/tablet?status=" + status.getText());
+            } else new Request().start("/tablet?status=" + status.getText());
             countDownTimer.start();
         }
     };
